games/libretro-craft: Import libretro-craft-20180803
Libretro is a simple but powerful development interface that allows for
the easy creation of emulators, games and multimedia applications that
can plug straight into any libretro-compatible frontend. This development
interface is open to others so that they can run these pluggable emulator
and game cores also in their own programs or devices.

Craft is a Minecraft clone. It's just a few thousand lines of C and uses
modern OpenGL (shaders).