libretro-fuse: add version 20180419.
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

FUSE is an emulator for the Sinclair ZX Spectrum and related machines.