libretro-beetle-wswan: add version 20180620.
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

This package contains a port of the Mednafen WonderSwan/WonderSwan Color
emulator (based on Cygne) for Libretro.