py-test: updated to 3.8.2
pytest 3.8.2:

Deprecations and Removals
* The item parameter of pytest_warning_captured hook is now documented as deprecated. We realized only after the 3.8 release that this parameter is incompatible with pytest-xdist.

Our policy is to not deprecate features during bugfix releases, but in this case we believe it makes sense as we are only documenting it as deprecated, without issuing warnings which might potentially break test suites. This will get the word out that hook implementers should not use this parameter at all.

In a future release item will always be None and will emit a proper warning when a hook implementation makes use of it.

Bug Fixes
* Fix reload on assertion rewritten modules.
* The .user_properties attribute of TestReport objects is a list of (name, value) tuples, but could sometimes be instantiated as a tuple of tuples. It is now always a list.
* No longer issue warnings about using pytest_plugins in non-top-level directories when using --pyargs: the current --pyargs mechanism is not reliable and might give false negatives.
* Exclude empty reports for passed tests when -rP option is used.
* Improve error message when an invalid Python expression is passed to the -m option.
* MonkeyPatch.setenv and MonkeyPatch.delenv issue a warning if the environment variable name is not str on Python 2.

In Python 2, adding unicode keys to os.environ causes problems with subprocess (and possible other modules), making this a subtle bug specially susceptible when used with from __future__ import unicode_literals.

Improved Documentation
* Add possible values for fixture scope to docs.